DashBoard/requirements.txt
streamlit==1.40.0
boto3==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
matplotlib==3.9.0
plotly==5.22.0
requests==2.31.0
